more-less/requirements.txt
fastapi[standard]
pyyaml